src/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
async-timeout==4.0.3; python_version < '3.11'
attrs==23.2.0; python_version >= '3.7'
cachetools==5.3.3; python_version >= '3.7'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
frozenlist==1.4.1; python_version >= '3.8'
google-auth==2.29.0; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
iso8601==2.1.0; python_version >= '3.7' and python_version < '4.0'
kopf==1.37.2; python_version >= '3.8'
kubernetes==29.0.0; python_version >= '3.6'
multidict==6.0.5; python_version >= '3.7'
oauthlib==3.2.2; python_version >= '3.6'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-json-logger==2.0.7; python_version >= '3.6'
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==2.0.0; python_version >= '3.4'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
typing-extensions==4.11.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
websocket-client==1.8.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
